Let the compiler type-check printf-style format strings
PrintFormattedCompat is a varargs function with no format attribute, so neither
clang nor gcc checked a single one of its ~200 call sites. A wrong conversion or
a missing argument is undefined behaviour that no test reliably catches, and a
grep cannot see a type mismatch at all.

Adding __attribute__((format(printf, 1, 2))) hands the whole class to the
compiler, permanently and on every platform, instead of to a scan that has to be
remembered and re-run.

It immediately found four sites passing a 64-bit size_type to %02d, in the two
flash-progress messages. All four now cast to int32_t, matching the cast the same
file already uses a few lines further down.

A separate scan for argument-count mismatches across all 200 call sites found
none, once it was taught that C++ concatenates adjacent string literals -- the
first version of it reported eleven false positives for exactly that reason, and
was only trusted after it was checked against planted cases.

+++ libraries/opensyde_core/protocol_drivers/basic_update/C_OscBuSequences.cpp
@@ -661,7 +661,7 @@ std::error_code C_OscBuSequences::UpdateNode(const std::string & orc_HexFilePath
                {
                   std::string c_Text;
                   c_Text = PrintFormattedCompat("Transferring area %02d/%02d  byte %08d/%08d",
-                                        u16_Area + 1, pc_HexDump->at_Blocks.size(),
+                                        u16_Area + 1, static_cast<int32_t>(pc_HexDump->at_Blocks.size()),
                                         s32_Size - s32_RemainingBytes, s32_Size);
                   u64_TotalSizeTransferred += c_Data.size();
                   m_ReportProgress(c_Return.value(), c_Text);
@@ -691,7 +691,8 @@ std::error_code C_OscBuSequences::UpdateNode(const std::string & orc_HexFilePath
          {
             std::string c_Text;
             c_Text = PrintFormattedCompat("Transferring area %02d/%02d  byte %08d/%08d",
-                                  u16_Area + 1, pc_HexDump->at_Blocks.size(), s32_Size, s32_Size);
+                                  u16_Area + 1, static_cast<int32_t>(pc_HexDump->at_Blocks.size()), s32_Size,
+                                  s32_Size);
 
             const uint8_t u8_MAX_PERCENTAGE = 100;
             m_ReportProgressPercentage(u8_MAX_PERCENTAGE, false);
